+++ rtlib/src/ray.rs
@@ -14,7 +14,7 @@ impl Ray {
         }
     }
 
-    pub fn position(&self, t: f32) -> Vec4 {
+    pub fn position(&self, t: f64) -> Vec4 {
         self.origin + (self.direction * t)
     }
